Laboratory work 2/requirements.txt
numpy>=1.20.0
Pillow>=8.0.0

